requirements.txt
Django>=4.2,<5.0  # نسخه پایدار Django
django-adminlte3==0.1.6  # آخرین نسخه پایدار
pandas==2.2.3
numpy==1.26.4
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.3.4
certifi==2024.8.30
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.0
django-axes==7.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.7.0
django-elasticsearch-dsl==8.0
django-encrypted-model-fields==0.6.5
django-fernet-fields==0.6
django-formtools==2.5.1
django-jazzmin==3.0.1
django-jet==1.0.8
django-material==1.12.0
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-timezone-field==7.1
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
elastic-transport==8.15.1
elasticsearch==8.16.0
elasticsearch-dsl==8.16.0
gunicorn==21.2.0
kombu==5.4.2
packaging==24.2
pillow==11.0.0
prompt_toolkit==3.0.50
pycparser==2.22
pypng==0.20220715.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
qrcode==7.4.2
redis==5.2.1
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
daphne>=4.0
django_select2
jdatetime
persiantools
channels
channels_redis
django-simple-history==3.1.1
WeasyPrint
flower==2.0.1
django-debug-toolbar
phonenumbers
ippanel
django_celery_results
django-cleanup
# ------------------- App: py_ai -------------------
#hazm==0.7.0
#transformers==4.41.1
#torch>=2.2.2
#scikit-learn==1.5.0
#matplotlib==3.9.0